pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "market-intel"
version = "0.1.0"
description = "Enterprise stock market education and analytics platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "lxml>=5.3.0",
    "numpy>=2.1.0",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "python-multipart>=0.0.20",
    "redis[hiredis]>=5.2.1",
    "sentencepiece>=0.2.0",
    "streamlit>=1.41.1",
    "tenacity>=9.0.0",
    "torch>=2.5.1",
    "transformers>=4.47.1",
    "uvicorn[standard]>=0.34.0",
    "websocket-client>=1.8.0",
    "websockets>=14.1",
    "yfinance>=0.2.51",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.4",
    "types-aiofiles>=24.1.0.20240626",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
